Renamed UpdateClient to CaseUpdateClient, update package to caseupdate
